+++ module/packer/signer_indices.go
@@ -2,8 +2,6 @@ package packer
 
 import (
 	"fmt"
-
-	"github.com/onflow/flow-go/consensus/hotstuff/model"
 )
 
 func DecodeSignerIndices(indices []byte, count int) ([]int, error) {
@@ -29,8 +27,7 @@ func DecodeSignerIndices(indices []byte, count int) ([]int, error) {
 	// remaining bits (if any), they must be all `0`s
 	remainings := byt << (8 - offset)
 	if remainings != byte(0) {
-		return nil, fmt.Errorf("the remaining bites are expected to be all 0s, but are %v: %w",
-			remainings, model.ErrInvalidFormat)
+		return nil, fmt.Errorf("the remaining bites are expected to be all 0s, but are %v", remainings)
 	}
 
 	return signerIndices, nil
